Split WASM wrapper tests by responsibility
Move the large wrapper test module into focused child modules for
approval, attachments, channel behaviour, cloning, conversion, dispatch,
and store coverage.

Keep the existing test bodies and expectations while reducing the
cohesion burden on a single test file.
